// UI/src/app/_shared/model/partComponent.ts
import { PartDocumentModel } from "./partdocumentmodel";

export class PartComponent {
  id: number;
  yazakipartnumber: string;
  tpartnumber: string;
  supplierpartnumber: string;
  customerpartnumber: string;
  Comodity: string;
  description: string;
  suppliername: string;
  customername: string;
  noofpoles: string;
  pinrows: string;
  sealed: string;
  gender: string;
  series: string;
  platingcolor: string;
  color: string;
  material: string;
  hybrid: string;
  width: string;
  length: string;
  height: string;
  uom: string;
  maxtemp: string;
  mintemp: string;
  od: string;
  applicablewirerange: string;
  wireinsulationdiarange: string;
  csa: string;
  noofstrands: string;
  wallthickness: string;
  sealdia: string;
  bundledia: string;
  wholesize: string;
  panelthickness: string;
  relatedpart: string;
  //matingpart: string;
  status: string;
  weight: string;
  grommethardness: string;
  wirestandard: string; 
  rank: string;
  remarks: string;
  seal: string;
  dummyplug: string;
  accessory: string;
  matingypn: string;
  matingsupplierpart: string;
  matingsuppliername: string;
  terminal: string;
  lastupdatedby: string;
  lastupdatedon: Date;
  action_On: any;
  action: string;
  action_By: string;  
  ldocumentlist: PartDocumentModel[] = [];
  FileDetailsModel() {
    this.id = null;
    this.yazakipartnumber = null;
    this.tpartnumber = null;
    this.supplierpartnumber = null;
    this.customerpartnumber = null;
    this.Comodity = null;
    this.description= null;
    this.suppliername= null;
    this.customername= null;
    this.noofpoles = null;
    this.pinrows = null;
    this.sealed = null;
    this.gender = null;
    this.series = null;
    this.platingcolor = null;
    this.color = null;
    this.material = null;
    this.hybrid = null;
    this.width = null;
    this.length = null;
    this.height = null;
    this.uom = null;
    this.maxtemp = null;
    this.mintemp = null;
    this.od = null;
    this.applicablewirerange = null;
    this.wireinsulationdiarange = null;
    this.csa = null;
    this.noofstrands = null;
    this.wallthickness = null;
    this.sealdia = null;
    this.bundledia = null;
    this.wholesize = null;
    this.panelthickness = null;
    this.relatedpart = null;
    //this.matingpart = null;
    this.status = null;
    this.weight = null;
    this.rank = null;
    this.remarks = null;
    this.seal = null;
    this.dummyplug = null;
    this.accessory = null;
    this.matingypn = null;
    this.matingsupplierpart = null;
    this.matingsuppliername = null;
    this.lastupdatedby = null;
    this.lastupdatedon = null;
    this.ldocumentlist = null;
  }
}