// UI/src/app/_shared/model/partdocumentmodel.ts
export class PartDocumentModel
{
  id: number;
  partnumber: string;
  date: string;
  doc_type: string;
  doc_name: string;
  doc_path: string;
  doc_image: string;
  errors: string;
  uploadedby: string;
  uploadedbyid: string;
  lastmodifieddate: string;
  lastmodifiedby: string;
  lastmodifiedbyid: string;  

  //new 

  
}